+++ MuZe/src/main/java/com/kh/muze/common/template/LoginUser.java
@@ -9,15 +9,13 @@ import com.kh.muze.member.model.vo.Member;
 
 @Component
 public class LoginUser {
-	
-	
 	public static int getUserNo(HttpSession session) {
-		
 		Member loginUser = (Member)session.getAttribute("loginUser");
 		int userNo = loginUser.getUserNo();
-		
 		return userNo;
 	}
-	
-
 }
+
+
+
+
