+++ MuZe/src/main/webapp/WEB-INF/views/diary/diaryCalender.jsp
@@ -257,7 +257,7 @@
             },
             // 이벤트 클릭시 다이어리 내용을 볼수 있는 이벤트
             eventClick : function(info) {
-            	// id가 999인건 일정 정보 아닌것은 다이어리
+            	// id가 999인건 일정 정보 , 아닌것은 다이어리
             	if(info.event.groupId != '999'){
 	            	$('#modal-content').modal('show');
 	            	$('#replyDiaryTitle').text(info.event.title);
@@ -444,7 +444,6 @@
         <div class="modal-dialog modal-lg">
             <div class="modal-content">
             <!-- USER_NO HIDDEN / DIARY_DATE HIDDEN -->
-            <input type="hidden" name="diaryUser" value="${sessionScope.loginUser.userNo}"/>
 			<input type="hidden" name="diaryDate" id="diaryDate"/>
                 <!-- Modal Header -->
                 <div class="modal-header" style="text-align: center;" id="diary-header">
@@ -524,12 +523,6 @@
                                               제목 <br>
                     <input type="text" placeholder="제목을 입력해세요..." name="diaryTitle" id="updateformTitle" class="diary-body" required> 
                    	<br/><br/>
-                   	<!-- 
-                    <div class="filebox">
-					    <input class="upload-name" value="첨부파일" placeholder="첨부파일" >
-					    <label for="file" id="file-box"  onclick="fileUploadClick();">파일찾기</label> 
-					</div>
-                   	 -->
                    		<input type="file" name="upfile" id="diary-file"> <br/><br/>
                     <div id="updateImage-area">
                     	<img src="" id="updateformImg"/>
